ml_service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
scikit-surprise==1.1.3
scikit-learn==1.4.2
pandas==2.2.2
numpy==1.26.4
pymongo==4.7.2
joblib==1.4.2
python-dotenv==1.0.1
pydantic==2.7.1
requests==2.31.0
apscheduler==3.10.4
